requirements.txt
fastapi==0.115.2
pydantic==2.9.2
SQLAlchemy==2.0.35
uvicorn==0.31.1
python-dotenv==1.0.1
asyncpg==0.29.0
aiohttp==3.11.11
APScheduler==3.11.0
psycopg2-binary==2.9.10
aiogram==3.17.0